Rename file with `ContextAssembler` to reflect its new content
